Tweaked Hyper parameter tuning
I made 3 minor modifications to the script. These are commented. 

If you want to get the cross-validation to run quickly just to check the code works, you can decrease the number of epochs and also decrease the size of the test data. I used 10 and 900, respectively, and the code completed fine in under 30s. Obviously you use the larger values when doing the proper run.